Añadir controlador de profesor y estilos a paginas junto con alertas

+++ Academico/app/Http/Controllers/MateriaController.php
@@ -11,7 +11,12 @@ class MateriaController extends Controller
     public function index()
     {
         $materias = Materia::all();
-        return view('materia.index', compact('materias'));
+        $breadcrumbs = [
+            ['url' => route('home'), 'name' => 'Inicio'],
+            ['name' => 'Materias', 'url' => '']
+        ];
+
+        return view('materia.index', compact('materias', 'breadcrumbs'));
     }
 
     public function show(Materia $materia)
@@ -22,7 +27,7 @@ class MateriaController extends Controller
     public function store(CreateMateriaRequest $request)
     {
         $request->createMateria();
-        return redirect()->route('materia.index');
+        return redirect()->route('materia.index')->with('success', 'Materia creado con exito');
     }
 
     public function create()
@@ -39,7 +44,7 @@ class MateriaController extends Controller
     public function update(UpdateMateriaRequest $request, Materia $materia)
     {
         $request->updateMateria($materia);
-        return redirect()->route('materia.index');
+        return redirect()->route('materia.index')->with('success', 'Aula editada con exito');
     }
 
     public function trash(Materia $materia)
@@ -51,14 +56,22 @@ class MateriaController extends Controller
 
     public function form($view, Materia $materia)
     {
-        return view($view, [
-            "materia" => $materia
-        ]);
+        $breadcrumbs = [
+            ['url' => route('home'), 'name' => 'Inicio'],
+            ['url' => route('materia.index'), 'name' => 'Materias'],
+        ];
+
+        if ($materia && $materia->exists) {
+            $breadcrumbs[] = ['name' => 'Editar Materia'];
+        } else {
+            $breadcrumbs[] = ['name' => 'Crear Materia'];
+        }
+        return view($view, compact('materia', 'breadcrumbs')); //["materia" => $materia]);
     }
 
     public function destroy($id)
     {
-        $materia = Materia::with('profesores')->find($id);
+        $materia = Materia::with('profesores', 'aulas')->find($id);
 
         if (!$materia) {
             return redirect()->route('materia.index')->with('error', 'La materia no existe.');
@@ -69,7 +82,12 @@ class MateriaController extends Controller
             return redirect()->route('materia.index')->with('error', 'Esta materia está asignada a un profesor y no puede ser eliminada.');
         }
 
-        // Si no está asignada a ningún profesor, procede a eliminar
+        if ($materia->aulas->isNotEmpty()) {
+            // Si la materia está asignada a alguna aula, redirige con un mensaje de error
+            return redirect()->route('materia.index')->with('error', 'Esta materia está asignada a una o más aulas y no puede ser eliminada.');
+        }
+
+        // Si no está asignada a ningún profesor ni aula, procede a eliminar
         $materia->delete();
         return redirect()->route('materia.index')->with('success', 'Materia eliminada con éxito.');
     }
